WIP: moves relevant components to /editor/interfaces; issues temporary import fixes
