+++ src/components/editor/controls/Channel/Row.jsx
@@ -1,6 +1,5 @@
 import styles from './Channel.css';
 import Step from './Step';
-import Sequencer from '@components/Sequencer/Sequencer.jsx';
 
 export default function Row({ notes, handleNoteChange }) {
   return (
