+++ src/components/editor/controls/index.js
@@ -2,7 +2,7 @@
 
 import Channel from './Channel/Channel.jsx';
 import Controls from './Channel/Controls.jsx';
-import Dropdown from '@components/Piano/DropDown.jsx';
+import Dropdown from './Channel/Dropdown.jsx';
 import Row from './Channel/Row.jsx';
 import Step from './Channel/Step.jsx';
 import GlobalControls from './GlobalControls/GlobalControls.jsx';
