+++ src/components/editor/interfaces/Piano/EffectsRack.jsx
