+++ src/components/editor/interfaces/Piano/Faders.jsx
