+++ src/components/editor/interfaces/Piano/Oscilloscope.jsx
