+++ src/components/editor/interfaces/Piano/Piano.css
