+++ src/components/editor/interfaces/Sequencer/Sequencer.css
