+++ src/components/editor/interfaces/index.js
@@ -1,1 +1,7 @@
 // handles exports of all components in current folder
+
+import DrumMachine from './DrumMachine/DrumMachine.jsx';
+import Piano from './Piano/Piano.jsx';
+import Sequencer from './Sequencer/Sequencer.jsx';
+
+export { DrumMachine, Piano, Sequencer };
